requirements.txt
Django==3.1.2
django-admin-rangefilter
python-decouple
dj-database-url
psycopg2-binary
gunicorn
whitenoise
django-storages
django-widget-tweaks
django-filter
vobject
humanize
python-telegram-bot
celery
django-celery-beat
django-celery-results
google-auth-oauthlib
google-api-python-client